+++ src/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -200,6 +200,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .eq("contribute", "contributing", "contribution").mayEnd(TXT),
             ofFile("crowdin", "/icons/crowdin.svg", "Crowdin: .crowdin.yml")
                 .eq(".crowdin.yml"),
+            ofFile("deptective", "/icons/deptective.png", "Deptective: deptective.json")
+                .eq("deptective.json"),
             ofFile("docker", "/icons/docker.png", "Docker: dockerfile(.xml,.yml,...)")
                 .eq("dockerfile").mayEnd(CFG),
             ofFile("dockercompose", "/icons/dockercompose.png", "Docker: docker-compose")