+++ src/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -361,6 +361,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 .eq("test", "testing").mayEnd(TXT),
             ofFile("todo", "/icons/todo.png", "To-Do: todo(.md,.txt,.adoc,.rst)")
                 .eq("todo").mayEnd(TXT),
+            ofFile("tox", "/icons/tox.png", "Tox: tox.ini")
+                .eq("tox.ini"),
             ofFile("travis", "/icons/travis.svg", "Travis CI: .travis.yml")
                 .eq(".travis.yml"),
             ofFile("vagrant", "/icons/vagrant.png", "Vagrant: vagrantfile")