+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.29.0 (WIP)
 * support `.teamcity` folder.
+* support `.exe` files.
 
 ### 1.28.0 (2020/05/23)
 * support Composer `composer.json` and `composer.lock` files.